Allow the first header to be different while using fast init
The fast initialization made available by supplying the number
of image file directories works by assuming the headers equally
spaced apart. If the first header file is different, which
sometimes is the case, the reader would be corrupted. Now
the 2nd and 3rd headers are used instead of the 1st and 2nd to
duplicate the rest of the header information. This allows the
first header to contain more information than the rest.